+++ volmdlr/core_compiled.pyx
@@ -335,6 +335,20 @@ class Vector(DessiaObject):
         except ZeroDivisionError:
             return False
 
+    def is_perpendicular_to(self, other_vector: "Vector", abs_tol: float = 1e-5):
+        """
+        Checks if two vectors are perpendicular.
+        The two vectors should be of same dimension.
+
+        :param other_vector: A vector-like object
+        :type other_vector: :class:`volmdlr.Vector`
+        :param abs_tol: Absolute tolerance to consider perpendicular
+        :type abs_tol: float
+        :return: `True` if the two vectors are perpendicular, `False` otherwise
+        :rtype: bool
+        """
+        return math.isclose(abs(self.dot(other_vector)), 0, abs_tol=abs_tol)
+
     @classmethod
     def mean_point(cls, points: List["Vector"]):
         """