+++ volmdlr/primitives3d.py
@@ -180,6 +180,9 @@ class Block(shells.ClosedShell3D):
                      self.frame.w.norm())
 
         faces = self.shell_faces()
+        for face in faces:
+            face.alpha = alpha
+            face.color = color
         shells.ClosedShell3D.__init__(self, faces, color=color, alpha=alpha, name=name)
 
     def __eq__(self, other):
@@ -511,7 +514,7 @@ class ExtrudedProfile(shells.ClosedShell3D):
         faces = self.shell_faces()
 
         shells.ClosedShell3D.__init__(self, faces, color=color,
-                                             alpha=alpha, name=name)
+                                      alpha=alpha, name=name)
 
     def to_dict(self, *args, **kwargs):
         """
@@ -555,7 +558,7 @@ class ExtrudedProfile(shells.ClosedShell3D):
             self.plane_origin, self.x, self.y)
         lower_face = volmdlr.faces.PlaneFace3D(
             lower_plane, surfaces.Surface2D(self.outer_contour2d,
-                                                 self.inner_contours2d))
+                                            self.inner_contours2d))
 
         upper_face = lower_face.translation(self.extrusion_vector)
         lateral_faces = []
@@ -706,7 +709,7 @@ class RevolvedProfile(shells.ClosedShell3D):
 
         faces = self.shell_faces()
         shells.ClosedShell3D.__init__(self, faces, color=color,
-                                             alpha=alpha, name=name)
+                                      alpha=alpha, name=name)
 
     def to_dict(self, *args, **kwargs):
         """
@@ -750,9 +753,9 @@ class RevolvedProfile(shells.ClosedShell3D):
             # Adding contours face to close
             w = self.x.cross(self.y)
             plane1 = surfaces.Plane3D(volmdlr.Frame3D(self.plane_origin,
-                                                           self.x,
-                                                           self.y,
-                                                           w))
+                                                      self.x,
+                                                      self.y,
+                                                      w))
             face1 = volmdlr.faces.PlaneFace3D(
                 plane1, surfaces.Surface2D(self.contour2d, []))
             face2 = face1.rotation(self.axis_point, self.axis, self.angle)
@@ -890,7 +893,7 @@ class RevolvedProfile(shells.ClosedShell3D):
         self.axis_point.frame_mapping_inplace(frame, side)
 
 
-class Cylinder(RevolvedProfile):
+class Cylinder(shells.ClosedShell3D):
     """
     Creates a full cylinder with the position, the axis of revolution the radius and the length.
     """
@@ -905,23 +908,25 @@ class Cylinder(RevolvedProfile):
         self.axis = axis
         self.radius = radius
         self.length = length
-        self.bounding_box = self._bounding_box()
+        frame_origin = position - axis * length * 0.5
+        self.frame = volmdlr.Frame3D.from_point_and_vector(frame_origin, axis, volmdlr.Z3D)
+        faces = self.shell_faces()
+        shells.ClosedShell3D.__init__(self, faces=faces, color=color, alpha=alpha, name=name)
 
-        # Revolved Profile
-        point1 = volmdlr.Point2D(-0.5 * self.length, 0.)
-        point2 = volmdlr.Point2D(0.5 * self.length, 0.)
-        point3 = volmdlr.Point2D(0.5 * self.length, self.radius)
-        point4 = volmdlr.Point2D(-0.5 * self.length, self.radius)
-        line_seg1 = volmdlr.edges.LineSegment2D(point1, point2)
-        line_seg2 = volmdlr.edges.LineSegment2D(point2, point3)
-        line_seg3 = volmdlr.edges.LineSegment2D(point3, point4)
-        line_seg4 = volmdlr.edges.LineSegment2D(point4, point1)
-        contour = volmdlr.wires.Contour2D([line_seg1, line_seg2, line_seg3, line_seg4])
-        y = axis.random_unit_normal_vector()
-        RevolvedProfile.__init__(self, position, axis, y, contour, position,
-                                 axis, color=color, alpha=alpha, name=name)
+    def shell_faces(self):
+        surface3d = surfaces.CylindricalSurface3D(self.frame, self.radius)
+        cylindrical_face = volmdlr.faces.CylindricalFace3D.from_surface_rectangular_cut(
+            surface3d, 0, 2*math.pi, 0, self.length)
+        lower_plane = surfaces.Plane3D.from_plane_vectors(
+            self.frame.origin, self.frame.u, self.frame.v)
+        lower_face = volmdlr.faces.PlaneFace3D(
+            lower_plane, surfaces.Surface2D(
+                volmdlr.wires.Circle2D(self.position.to_2d(self.frame.origin, self.frame.u,
+                                                           self.frame.v), self.radius), []))
+        upper_face = lower_face.translation(self.frame.w * self.length)
+        return [lower_face, cylindrical_face, upper_face]
 
-    def _bounding_box(self):
+    def get_bounding_box(self):
         """
         Computes the bounding box of a cylinder.
 
@@ -1651,7 +1656,7 @@ class Sweep(shells.ClosedShell3D):
 
         faces = self.shell_faces()
         shells.ClosedShell3D.__init__(self, faces, color=color,
-                                             alpha=alpha, name=name)
+                                      alpha=alpha, name=name)
 
     def to_dict(self, *args, **kwargs):
         """Custom serialization for performance."""
@@ -1755,10 +1760,10 @@ class Sweep(shells.ClosedShell3D):
                     points_3d.append(poly.points[0])
 
                 bezier_surface3d = surfaces.BezierSurface3D(degree_u,
-                                                                 degree_v,
-                                                                 points_3d,
-                                                                 size_u,
-                                                                 size_v)
+                                                            degree_v,
+                                                            points_3d,
+                                                            size_u,
+                                                            size_v)
 
                 outer_contour = volmdlr.wires.Contour2D([volmdlr.edges.LineSegment2D(volmdlr.O2D, volmdlr.X2D),
                                                          volmdlr.edges.LineSegment2D(